Refactor: Move handleMultiItemUnsavedChanges to dedicated MultiEditUnsavedChangesService
Move the multi-edit unsaved changes orchestration logic from the UI components
layer to the authoring services layer for better semantic organization.

Changes:
- Create new MultiEditUnsavedChangesService.ts in authoring/services
- Move handleMultiItemUnsavedChanges function to the new service
- Update imports in multiedit.ts and multi-edit-modal.tsx
- Add backward-compatible re-export in prompt-for-unsaved-changes.tsx

This improves code organization by separating business logic (services) from
UI components, making the codebase more maintainable and easier to discover
multi-edit utilities.
